Move icons to src
Assets in public directory cannot be imported from JavaScript.
If you intend to import that asset, put the file in the src directory,
and use /src/icons/filter.svg?react instead of
/public/icons/filter.svg?react.
If you intend to use the URL of that asset, use
/icons/filter.svg?url&react.
