+++ src/client/src/commons/menu/mainView/mainSideNav.tsx
@@ -3,12 +3,12 @@ import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { Badge, Stack } from "@mui/material";
 import { ImportErrorModal } from "./menuComponents/importErrorModal";
-import FilterIcon from "../../../../public/icons/filter.svg?react";
-import AddIcon from "../../../../public/icons/add.svg?react";
-import UploadIcon from "../../../../public/icons/upload.svg?react";
-import SettingsIcon from "../../../../public/icons/settings.svg?react";
-import HelpIcon from "../../../../public/icons/help.svg?react";
-import LayersIcon from "../../../../public/icons/layers.svg?react";
+import FilterIcon from "../../../assets/icons/filter.svg?react";
+import AddIcon from "../../../assets/icons/add.svg?react";
+import UploadIcon from "../../../assets/icons/upload.svg?react";
+import SettingsIcon from "../../../assets/icons/settings.svg?react";
+import HelpIcon from "../../../assets/icons/help.svg?react";
+import LayersIcon from "../../../assets/icons/layers.svg?react";
 import { theme } from "../../../AppTheme";
 import ImportModal from "./actions/importModal";
 import { DrawerContentTypes } from "../../../pages/editor/editorComponentInterfaces";
