+++ src/client/src/commons/menu/profilePopup.tsx
@@ -1,7 +1,7 @@
 import { useAuth } from "react-oidc-context";
 import { Button, IconButton, Popover, Stack, Typography } from "@mui/material";
 import { theme } from "../../AppTheme.ts";
-import ProfileIcon from "../../../public/icons/profile.svg?react";
+import ProfileIcon from "../../assets/icons/profile.svg?react";
 import { UserData } from "../../ReduxStateInterfaces";
 import { MouseEvent, useState } from "react";
